Submit exercises 4.18 - 4.23 solution

+++ part4/blog-list/controllers/blogs.js
@@ -3,9 +3,9 @@ const blogsRouter = express.Router()
 const Blog = require('../models/blog.js')
 const User = require('../models/user.js')
 
-// Get all blogs
+// Get all blogs related to a user
 const getAllBlogs = async (req, res) => {
-  const blogs = await Blog.find({}).populate('user', {
+  const blogs = await Blog.find({ user: req.user.id }).populate('user', {
     username: 1,
     name: 1,
   })
@@ -15,14 +15,15 @@ const getAllBlogs = async (req, res) => {
 // Create a blog
 const createBlog = async (req, res, next) => {
   const data = req.body
-  const newBlog = new Blog(data)
+  const newBlog = new Blog({ ...data, user: req.user.id })
 
   try {
     const savedBlog = await newBlog.save()
 
-    const users = await User.findById(savedBlog.user)
-    users.blogs = users.blogs.concat(savedBlog.id)
-    await users.save()
+    // We will assign user to the req after a successful token verification
+    const user = await User.findById(req.user.id)
+    user.blogs = user.blogs.concat(savedBlog.id)
+    await user.save()
 
     res.status(201).json(savedBlog)
   } catch (error) {
@@ -58,8 +59,15 @@ const deleteBlog = async (req, res, next) => {
   const { id } = req.params
 
   try {
-    await Blog.findByIdAndRemove(id)
-    res.status(204).end()
+    const blog = await Blog.findById(id)
+    if (blog.user.toString() === req.user.id) {
+      await Blog.findByIdAndRemove(id)
+      res.status(204).end()
+    } else {
+      res
+        .status(401)
+        .json({ error: 'User is not authorized to perform the operation' })
+    }
   } catch (error) {
     next(error)
   }
